Move ActiveRecord::ConnectionAdapters::ConnectionHandler to separate file
